Add files via upload
The public diabetes dataset is in the csv file. We read this file in the Jupyter Notebook. 
The .ipynb file is the Jupyter Notebook with the code we used to explore our data. Reminder that this is not an all inclusive script. It is just the beginning to what you can do with pandas and python in terms of analyzing data. 
The .html file is the pandas profiling report generated in the Jupyter Notebook.